Log rotation, no (crashing) attempt to log during init.
Logs are rotated when they are at least 1MB, on startup, and immediately
before and after a dump. (Never in the middle of a dump.)

Note that as long as old clients are connected, they will continue to
write to whatever log they started out with. This should probably be
fixed to only apply to single transactions.