Add three (as yet unused) fields to post_t.
modified is the date some data relating to this image was last modified.
(Not tagnames or other data not directly related to the image. Not intended
to reflect tag-changes that result from soft implications either, if I get
around to implementing those.)

image_date is when the image was produced.
image_date_fuzz is how exact this date is. (Image was made within
image_date +- image_date_fuzz. I realize this needs to be bigger than 16
bits or have a non-linear scale. Maybe something like this?:
0-59: seconds
-=60
0-59: minutes
-=60
0-23: hours
-=24
0-365: days
-=366
0-65025: years
Or maybe a bit less on years, with millenia after..
)