Move log-dumps into their own directory (with their own sequence numbers).
The semantics are still broken though. (No modifications can be allowed
during dumping (which could be achieved with just mm_lock()), the normal
log needs to be rotated immedietaly after, so there can be no ongoing
transactions (currently not possible to ensure, and no rotation happens),
and which log to start from after a dump needs to be recorded.

And of course the code to recover without the cache doesn't fully exists
(only for loading a dump).