Remove server functionality from pgtest.
So now it's "./pgtest" to populate mm_cache (fails if there's anything
there), and then "./server" to serve it. Currently not possible to read
data from a log, but of course server will do that automatically
eventually.

pgtest should probably dump a log when it's done, but currently there's
no need to wait for that.