Mess with mm_init.
use_existing no longer implies failure if the existing cache is not there.
Soon I will change it to check the cache better, and only succeed if it
looks like a clean shutdown. (So it can always med called with
use_existing, and repopulate from log if it fails. (Should probably invert
the return value as well.)