Get rid of the flock() nonsense.
It was intended to keep concurrent processes from stepping on each other.
Since the memory management doesn't actually work with several processes
anyway, all it does is add overhead and hinder portability.

I'm leaving the calls to trans_(un)lock(), they might be useful as a
reminder that the server could become threaded in the future. (The cache
walker doesn't count.)