Expose image date/created in magic tags. (Disabled.)
I'm leaving this undocumented (and it's not enabled automatically), because
the tags are just set initially from the old value, then they stop being
connected. I actually want to remove the old version completely, but maybe
not at the protocol level.

At any rate, I consided this half baked, and actually using it might leave
you with logs that aren't valid on a newer server.