Start requiring fixup for datatags. New log version.
Some properties can no longer be set except in old logs (ratings are gone,
and source/title/score move to being tags only).

I still need to not return them as properties, and dates need a new format.
I won't bother to handle logs produced by this version when this is fixed,
so don't run it.

I'm also not commiting the documentation on how to do the fixup yet, as an
extra indication that this version is not to be used.