Tagvalue strings have to live in mm, even when constant.
I could to some kind of hack to move the valuelist to mm, but it's not
enough memory to worry about. Other parts of tagvalues leak more than this.
(A simple refcounting reuse of mm-strings might be worthwile. Or even
without refcounting. Several other objects also just stay around when they
are no longer used.)