Normalize utf in string-values.
This is a pretty big diff mostly because I had to change a lot of
prototypes. I'm reusing cmd (part of the decoded linebuf for the
connection) to store the recoded values (they are composed, and as such
never longer than the original), and passing that around as writeable
required this change.

The other alternative was worrying about where allocated memory gets freed,
and that didn't feel particularly appealing. (Searches would need a cleanup
fuction, and you can't free a const pointer, so it's basically the same
problem all over again, just in different places.)

So in summary, manual memory management is love.