General support for one or two direction fuzz.
This isn't parsed properly in the datetime case, but other than that it
might work. The point of course is that a two direction fuzz can't express
an even interval length. Plus specifying it in one direction only is
sometimes easier.

So now the implied fuzz of a day (etc) is actually accurate, that is it
stretches from midnight until 23:59:59. (So adjacent days don't overlap,
but one of them being +-1S is enough to match them.)

Note also that this changes the meaning of old logs, if you actually used
fuzz for anything on ints or uints (as the fuzz is now signed, and
therefore base 10). I haven't used this, and you probably haven't either.
But in case you have, you'll have to edit your old logs. Sorry about that.