Implications should not override explicit values.
It's been a long time since I looked at this code, so I'm not all that
confident that this fix is correct. But the hope is that if a implies b
with inherited value and you set both a and b on a post, b will retain the
value you set. (And previously it would get the value you set on a.)